Make changes according to code review
