--- src/anbox/dbus/SensorsServer.cpp
@@ -1,29 +1,0 @@
-/*
- * Copyright (C) 2020 Tomasz Grobelny <tomasz@grobelny.net>
- *
- * This program is free software: you can redistribute it and/or modify it
- * under the terms of the GNU General Public License version 3, as published
- * by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranties of
- * MERCHANTABILITY, SATISFACTORY QUALITY, or FITNESS FOR A PARTICULAR
- * PURPOSE.  See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-#include "anbox/dbus/SensorsServer.h"
-
-#include "anbox/android/intent.h"
-#include "anbox/logger.h"
-
-double SensorsServer::Temperature() {
-  return impl_->temperature;
-}
-
-void SensorsServer::Temperature(const double& value) {
-  impl_->temperature = value;
-}
