+++ src/anbox/dbus/application_manager.xml
