+++ src/anbox/dbus/sensors.xml
